fix bug when user id is too big

+++ djang/bot/models.py
@@ -4,7 +4,7 @@ from django.utils import timezone
 
 
 class User(models.Model):
-    id = models.IntegerField(primary_key=True)
+    id = models.BigIntegerField(primary_key=True)
     username = models.CharField(max_length=30)
     first_anme = models.CharField(max_length=30)
     last_name = models.CharField(max_length=30)
